Move logging tests to it package and start them with failsafe plugin in their own VM
